consistent namespaces usage
